Add files via upload
Exercice 1 : List the 15 cities with the most airports in the database to obtain the result sorted in descending order.

Exercice 2 : Build and display the number of airlines per country. Draw points on the map proportional to the number of airlines per country. We can use the coord_country dataframe defined in the file provided.